pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dwarf-alpaca"
version = "0.1.0"
description = "ASCOM Alpaca server exposing a DWARF 3 telescope as Telescope/0, Camera/0, and Focuser/0 devices."
readme = "README.md"
authors = [
  { name = "Astro Tools" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "alpyca>=1.0.2",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "structlog>=24.1.0",
  "bleak>=0.21.0",
  "aiortc>=1.6.0",
  "av>=10.0.0",
  "numpy>=1.26.0",
  "opencv-python-headless>=4.9.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "protobuf>=3.20.3,<3.21",
  "websockets>=12.0"
]

[project.optional-dependencies]
development = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "httpx[cli]>=0.27.0",
  "ruff>=0.5.5"
]

[project.scripts]
dwarf-alpaca = "dwarf_alpaca.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dwarf_alpaca"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:google.protobuf.internal.well_known_types",
]
